+++ README.md
@@ -29,4 +29,4 @@
 #### 요구사항
 - [x] ThreadPoolExecutor를 적용해 일정 수의 사용자 동시에 처리가 가능하도록 한다.
 - [x] HTTP 응답 Header/Body 처리만을 담당하는 역할을 분리
-- [ ] Servlet을 Controller로 변경
+- [x] Servlet을 Controller로 변경
