+++ src/main/java/utils/ResourceMatcher.java
@@ -1,4 +1,4 @@
-package web.filter;
+package utils;
 
 import web.ContentType;
