--- src/main/java/web/filter/Filter.java
@@ -1,11 +1,0 @@
-package web.filter;
-
-import web.HttpRequest;
-import web.HttpResponse;
-
-import java.io.IOException;
-
-public interface Filter {
-
-    void doFilter(HttpRequest httpRequest, HttpResponse httpResponse, FilterChain filterChain) throws IOException;
-}
