+++ src/main/java/web/servlet/ResourceController.java
@@ -5,7 +5,7 @@ import web.HeaderName;
 import web.HttpRequest;
 import web.HttpResponse;
 import web.HttpStatusCode;
-import web.filter.ResourceMatcher;
+import utils.ResourceMatcher;
 
 import java.io.IOException;
